+++ config/config.go
@@ -183,7 +183,7 @@ type MainConfig struct {
 	Filters        []string               `json:"filters"`
 	MetricRenames  []MetricRenamesConfig  `json:"metric_renames"`
 	StaticMetadata []StaticMetadataConfig `json:"static_metadata"`
-	LogConfig      LogConfig              `json:"log_config"`
+	LogConfig      LogConfig              `json:"log"`
 
 	DisableSupervisor  bool `json:"disable_supervisor"`
 	DisableDiagnostics bool `json:"disable_diagnostics"`
@@ -284,7 +284,7 @@ func Configure(args []string, readFunc FileReadFunc) (MainConfig, map[string]str
 		a.Flag(lowerPrefix+".timeout", upperPrefix+" timeout used for OTLP Export() requests").
 			DurationVar(&op.Timeout.Duration)
 
-		a.Flag(lowerPrefix+".compression", upperPrefix+" compression used for OTLP requests (e.g., snappy).").
+		a.Flag(lowerPrefix+".compression", upperPrefix+" compression used for OTLP requests (e.g., snappy, gzip, none).").
 			StringVar(&op.Compression)
 	}