+++ CHANGELOG.md
@@ -16,10 +16,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
   option: `max_timeseries_per_request`. (#128)
 - The sidecar's max shards is now configurable via `--prometheus.max-shards`. There
   is also a matching yaml configuration option: `max_shards`. (#128)
+
 ### Changed
 - The sidecar's WAL-reader addresses several race conditions by monitoring
   Prometheus for readiness and the current segment number during WAL segment
   transitions. (#118)
+
 ### Removed
 - Remove the use_meta_labels parameter. (#125)